feat: method for listing ids of chunks from file data
This will be useful soon, to enable us to check how many chunks from
file data there is in the repository, to check that an upcoming chunk
size setting works. Also add an API call for returning the ids.

Note that all of this is meant for testing only. It may be best to
disable it in production builds, eventually.