fix: in an incremental backup, update total file count if needed
Previously, we were showing "N/M", where N is the number of files
found in the current backup run, and M the number of files in the
previous backup. I found this confusing if the new run finds more
files than were in the previous run. Now we increment M if N goes
beyond it.