build: don't hid clippy warnings
Previously we were hiding (with chronic) clippy output if it was
successful. However, clippy exits with zero even if it writes
warnings. Thus, chronic is the wrong solution. The output also
contains progress reporting from cargo, which is unwanted, thus, this
commit replaces chronic with the clippy -q option, to get rid of the
progress reporting.

Sponsored-by: author