fix(readme.md): link to rendered versions of other docs
Fixes #124.

I chose not to change the names of the link, they still say ".md". My
hope is that a person reading the Markdown document will understand that
we mean the .md document in this repo, and the person reading HTML will
have a valid link to doc.obnam.org.