fix: use an explicit tokio runtime
Create a tokio runtime so that the non-async function for chunkify can
call an async function and have that create and run background tasks.

Brown paper bag mistake: I did the development of this in a separate
program with an async main function. When I integrated it into Obnam
proper, which doesn't have an async main function (yet), I didn't
actually test it works, only that it compiles. Stupid, stupid, stupid.
I know better.

Sponsored-by: author