refactor: call NascentGeneration::insert from ::insert_iter
This is a step towards getting rid of insert_iter entirely, which
would make it easier to make `obnam backup` use async.

I originally split insert_iter so I could use a single transaction for
inserting many rows, but it seems to not be needed for speed after
all. I've benchmarked backing up a large file with and without this
change, and there's no real difference. I've not benchmarked with a
large number of files.

Even if there's a performance hit from using multiple transactions, my
hope is that by being able to use more CPUs/threads for backing up
will outweigh that by far.

Sponsored-by: author