test: tell Subplot-generated test program where the binaries are
Previously, we blindly assumed that Cargo puts binaries in
`target/debug` in the source tree. That is the default, but the user
can change it. Now we get the Cargo target directory by parsing the
output of "cargo metadata" and pass in the path to the test program
via an environment variable.

Also, reformat shell script.

Sponsored-by: author