feat! store schema version of generation database in the db
Add a new mandatory database table "meta" to the SQLite database the
stores information about the files in a backup generation. The idea is
for future versions of the Obnam client to be able to be able to
restore from backups made by older -- or newer -- versions of Obnam,
as far as is reasonable.

Add the `obnam gen-info` command to show information about the
generation metadata.

Sponsored-by: author