chore: add missing capture types for Subplot
Subplot now checks that all embedded files in the markdown file are
actually used by scenarios. To do this, it needs the scenario step
bindings to declare that a captured part of the scenario step is the
name of an embedded file. Add the couple of such type declarations
that were missing in Obnam.

Sponsored-by: author