feat! rename metadata field "sha256" to "label"
The field still contains a cleartext SHa256 of the cleartext chunk
data, but this makes it clearer that it may contain other data.

This is a breaking change: the server API won't work with an old
client, and the new client won't work with an old server. To avoid the
breakage would require more effort than is warranted at this time,
given the very small number of users of Obnam. Sorry.

Sponsored-by: author