refactor: add a type for plain integers we store in a database
This will make it easier to change later, if need be. We may want to
do that for various reasons, such as to save space. We may also want
to change things to only use integer types that SQLite can handle: u64
is currently not well handled by our database layer. However, as this
is a refactor, there's no change or fix to that.

FileId is now explicitly a database integer. This doesn't break
anything, for now, as the underlying integer type is still u64.

Also, change a couple of places where it will matter if DbInt changes
away from u64, and disable warnings for harmless conversions that
may cause warnings depending on what type DbInt has.

Sponsored-by: author