fix: use separate subplot YAML metadata file for new Subplot
Subplot has just made a breaking change: the subplot document metadata
must be in a separate YAML file, not embedded in the markdown. This
adapts Obnam to that.

It's a little awkward that Subplot hasn't made a release of this yet,
but as I use Subplot from git, I need this now.

Sponsored-by: author